requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymysql==1.1.0
cryptography==41.0.7
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
